requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
confusable-homoglyphs==3.2.0
cryptography==2.4.1
defusedxml==0.5.0
Django==2.1.2
django-adminlte2==0.4.1
django-allauth==0.38.0
django-registration==3.0
django-registration-redux==2.4
djangorestframework==3.9.0
idna==2.7
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.6.1
pycparser==2.19
PyJWT==1.6.4
PyMySQL==0.9.2
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
urllib3==1.24.1
